chore: add script to download and install tmtop latest release (#6)
Added a bash script to automate the download and installation of the latest release of tmtop for Linux AMD64. The script uses curl to fetch the latest release information from GitHub's API, extracts the download URL for the linux_amd64.tar.gz file, downloads it using wget, unpacks it, and then moves the executable to /usr/local/bin for easy execution.